backend/backend/requirements.txt
flask
pyzbar
pillow
gunicorn
qrcode
opencv-python
requests
